refactor(docs): split the four giant landing components
